requirements.txt
flask
flask-restful
flask-bootstrap
pandas
pymysql
pymongo
Pillow
requests
torchvision